fix(fpr): config/URL env vars are not credential taint sources (segment A source-precision)
